requirement.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.5.0
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.3
attrs==22.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.11.2
bleach==4.1.0
blis==0.7.9
Brotli==1.0.9
cached-property==1.5.2
cachetools==4.2.2
catalogue==1.0.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cinemagoer==2023.5.1
click==8.1.7
cohere==4.21
colorama==0.4.6
comm==0.1.2
confection==0.1.1
cymem==2.0.6
dataclasses==0.8
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
en-core-web-sm==2.3.1
entrypoints==0.4
executing==0.8.3
fastavro==1.8.2
fastjsonschema==2.16.2
filelock==3.12.3
frozenlist==1.4.0
fsspec==2023.6.0
greenlet==2.0.2
huggingface-hub==0.16.4
idna==3.4
IMDbPY==2022.7.9
importlib-metadata==6.8.0
ipykernel==6.25.0
ipython==8.12.2
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==2.11.3
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.3.0
jupyter-server==1.13.5
langcodes==3.3.0
llvmlite==0.40.0
lxml==4.9.3
MarkupSafe==1.1.1
matplotlib-inline==0.1.6
mistune==0.8.4
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==8.14.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.7
nbclassic==0.5.5
nbconvert==5.5.0
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.2
numba==0.57.0
numpy==1.24.3
openai==0.27.10
ordered-set==3.1.1
ordered-set-stubs==0.1.3
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
pickleshare==0.7.5
Pillow==10.0.0
pip==23.2.1
plac==1.1.3
platformdirs==3.10.0
poetry-version==0.1.5
preshed==3.0.6
profanity-filter==1.3.0
prometheus-client==0.14.1
prompt-toolkit==3.0.36
psutil==5.9.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.8
Pygments==2.15.1
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3
pywin32==305.1
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==23.2.0
redis==3.5.3
regex==2023.8.8
requests==2.31.0
ruamel.yaml==0.15.100
safetensors==0.3.3
Send2Trash==1
